CarpetReduce: Explicitly convert from/to CCTK_INT and from/to CCTK_REAL
Explicitly convert between int and CCTK_INT, and between double and
CCTK_REAL, where necessary.

darcs-hash:20060613171604-dae7b-6cdecab56f7600ef471621e808847d32e45baa75.gz

+++ Carpet/CarpetReduce/src/reduce.cc
@@ -988,7 +988,8 @@ namespace CarpetReduce {
             } else if (ilen >= 0) {
               assert (interp_num_time_levels>0);
               num_tl
-                = min (prolongation_order_time + 1, interp_num_time_levels);
+                = min (prolongation_order_time + 1,
+                       (int)interp_num_time_levels);
             } else {
               assert (0);
             }
